restore non_exhaustive to response-only models
